Es können jetzt 32-Bit und 64-Bit Versionen des Programms erstellt werden. Sämtliche double variablen in floats umgewandelt um Speicher zu sparen. CConnectionList verwendet jetzt statt einer verketteten Liste ein Array um Speicher zu sparen. CTSPPointList Neue Klasse CMemoryTester eingefügt um zu ermitteln ob der Speicher des Systems ausreicht um ein TSP zu laden.

+++ WindowsFormsApplication1/WindowsFormsApplication1/CAntAlgorithmParameters.cs
@@ -7,11 +7,11 @@ namespace WindowsFormsApplication1
 {
     public class CAntAlgorithmParameters
     {
-        public double initialPheromone;
-        public double pheromoneUpdate;
-        public double evaporationFactor;
-        public double pheromoneParameter;
-        public double localInformation;
+        public float initialPheromone;
+        public float pheromoneUpdate;
+        public float evaporationFactor;
+        public float pheromoneParameter;
+        public float localInformation;
         public int numberAnts;
         public int numberMaxIterations;
         public CTour optTour;
